renames home_screen to stations_screen
